Correct mixed up test names
