+++ test/unit/Quarter.UnitTest/State/ConfirmArchiveProjectActionTest.cs
@@ -1,11 +1,12 @@
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using NUnit.Framework;
-using Quarter.Components.Modals;
+using Quarter.Core.Commands;
 using Quarter.Core.Models;
 using Quarter.State;
-using Quarter.UnitTest.TestUtils;
+using Quarter.State.ViewModels;
 
 namespace Quarter.UnitTest.State;
 
@@ -19,32 +20,25 @@ public class ConfirmArchiveProjectActionTest : ActionHandlerTestCase
     {
         _state = NewState();
         _projectId = IdOf<Project>.Random();
-        _state.Modals.Push(ModalState.ParameterLess(typeof(FakeModal)));
-        _state = await ActionHandler.HandleAsync(NewState(), new ShowArchiveProjectAction(_projectId), CancellationToken.None);
+        _state.Projects = new List<ProjectViewModel> { new() { Id = _projectId, Name = "Alpha" } };
+        _state = await ActionHandler.HandleAsync(_state, new ConfirmArchiveProjectAction(_projectId), CancellationToken.None);
     }
 
     [Test]
-    public void ItShouldPushNewModal()
-        => _state.AssertPushedNewModal(typeof(ConfirmModal));
-
-    [Test]
-    public void ItShouldUseCorrectParameters()
+    public void ItShouldIssueCommand()
     {
-        var parameters = _state.Modals.Select(m => m.Parameters).First();
-        Assert.Multiple(() =>
-        {
-            Assert.That(parameters["Title"], Is.EqualTo("Archive project?"));
-            Assert.That(parameters["Message"], Is.EqualTo("If you archive this project it can no longer be used to register time. All registered time will still be available though. This project can be restored at a later time."));
-            Assert.That(parameters["ConfirmText"], Is.EqualTo("Archive"));
-        });
+        var expectedCommand = new ArchiveProjectCommand(_projectId);
+        AssertIssuedCommandByUserId(expectedCommand, ActingUserId);
     }
 
     [Test]
-    public void ItShouldIssueExpectedAction()
+    public void ItShouldBeMarkedAsArchivedInState()
     {
-        var parameters = _state.Modals.Select(m => m.Parameters).First();
-        var action = parameters[nameof(ConfirmModal.OnConfirmAction)] as ConfirmArchiveProjectAction;
-
-        Assert.That(action?.ProjectId, Is.EqualTo(_projectId));
+        var project = _state.Projects.First(p => p.Id == _projectId);
+        Assert.That(project.IsArchived, Is.True);
     }
+
+    [Test]
+    public void ItShouldPopTopMostModal()
+        => Assert.That(_state.Modals, Is.Empty);
 }
